+++ indigo_app/forms/__init__.py
@@ -1,3 +1,4 @@
+from .mixins import *
 from .documents import *
 from .places import *
 from .tasks import *
